+++ tests/test_cache.py
@@ -2,7 +2,6 @@
 # Copyright 2022-25 Board of Regents of the University of Wisconsin System
 # SPDX-License-Identifier: MIT
 
-import sys
 import time
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
@@ -44,43 +43,6 @@ def test_key_tmpdir():
     assert f(Path("a"), ("b",)) == Path("a/v0/fetching/b")
 
 
-def test_swap_in_fetched_file(tmp_path):
-    dest = tmp_path / "a" / "b" / "target"
-    src = tmp_path / "tmpfile"
-    src.touch()
-    _cache._swap_in_fetched_file(dest, src, timeout=0)
-    assert dest.is_file()
-    assert not src.is_file()
-
-    # With dest existing
-    src.touch()
-    _cache._swap_in_fetched_file(dest, src, timeout=0)
-    assert dest.is_file()
-    assert not src.is_file()
-
-    # With dest initially open
-    exec = ThreadPoolExecutor()
-
-    def close_after_delay(f):
-        time.sleep(0.1)
-        f.close()
-
-    src.touch()
-    with open(dest) as fp:
-        exec.submit(close_after_delay, fp)
-        _cache._swap_in_fetched_file(dest, src, timeout=10)
-        assert dest.is_file()
-        assert not src.is_file()
-
-    exec.shutdown()
-
-    # With dest left open
-    if sys.platform == "win32":
-        src.touch()
-        with open(dest) as fp, pytest.raises(OSError):
-            _cache._swap_in_fetched_file(dest, src, timeout=0.1)
-
-
 def test_move_in_fetched_directory(tmp_path):
     dest = tmp_path / "a" / "b" / "target"
     src = tmp_path / "tmpdir"
@@ -106,12 +68,3 @@ def test_wait_for_dir_to_vanish(tmp_path):
     path.mkdir()
     with pytest.raises(Exception):
         _cache._wait_for_dir_to_vanish(path, 0.1)
-
-
-def test_backoff_seconds():
-    f = _cache._backoff_seconds
-    assert list(f(1, 1, 0)) == [-1]
-    assert list(f(1, 1, 1)) == [1, -1]
-    assert list(f(1, 1, 0.1)) == [0.1, -1]
-    assert list(f(1, 5, 10, factor=2)) == [1, 2, 4, 3, -1]
-    assert list(f(1, 2, 10, factor=2)) == [1, 2, 2, 2, 2, 1, -1]